👽️ Socket.IO >> More cleanup and reorganizing.
